pyproject.toml
[project]
name = "fiber-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "apscheduler>=3.11.1",
    "asyncpg>=0.30.0",
    "asyncpg-lite>=0.3.1.3",
    "gevent>=25.9.1",
    "psycopg[binary]>=3.2.12",
    "python-decouple>=3.8",
]
